Split density graph settings from stream settings.
